Split truthiness analysis logic into dedicated files
The truthiness migration logic has grown sufficiently
to benefit from being split into separate files,
so let's do so.
