+++ src/analyzer/lib.rs
@@ -15,7 +15,7 @@ pub mod scope_analyzer;
 pub mod statements_analyzer;
 mod stmt;
 mod stmt_analyzer;
-mod truthiness_analyzer;
+mod truthiness;
 
 pub mod dataflow;
 pub mod function_analysis_data;
